Update README.md
Wednesday - What is the meaning of Wednesday Addams?
When the characters were adapted to the 1964 television series, Charles Addams gave her the name "Wednesday", based on the nursery rhyme line, "Wednesday's child is full of woe". The idea for the name was supplied by the actress and poet Joan Blake, an acquaintance of Addams.
Stranger Things - When a young boy vanishes, a small town uncovers a mystery involving secret experiments, terrifying supernatural forces and one strange little girl. Watch all you want. This nostalgic nod to '80s sci-fi and horror classics has earned dozens of Emmy nominations, including three for Outstanding Drama.
Enola Holmes - In the film, Enola travels to London to find her missing mother but ends up on a thrilling adventure, pairing up with a runaway lord as they attempt to solve a mystery that threatens the entire country. In addition to Brown, the film also stars Sam Claflin, Henry Cavill, and Helena Bonham Carter.
Elite - When three working-class teens enroll in an exclusive private school in Spain, the clash between them and the wealthy students leads to murder.